requirements.txt
blinker           1.9.0
click             8.1.8
colorama          0.4.6
exceptiongroup    1.2.2
Flask             3.1.0
Flask-Login       0.6.3
Flask-Mail        0.10.0
Flask-SQLAlchemy  3.1.1
Flask-WTF         1.2.2
greenlet          3.1.1
iniconfig         2.0.0
itsdangerous      2.2.0
Jinja2            3.1.5
MarkupSafe        3.0.2
packaging         24.2
pip               21.2.4
pluggy            1.5.0
PyMySQL           1.1.1
pytest            8.3.4
setuptools        58.1.0
SQLAlchemy        2.0.38
tomli             2.2.1
typing_extensions 4.12.2
Werkzeug          3.1.3
WTForms           3.2.1